requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
SQLAlchemy==2.0.44
sqlalchemy-utils==0.41.2
alembic==1.17.2
Flask==3.1.2
pydantic==2.12.4
python-dotenv==1.2.1
psycopg2-binary==2.9.11
pytest==9.0.1
pytest-asyncio==1.3.0
httpx==0.28.1
python-multipart==0.0.9
